requirements.txt
speechrecognition
pyttsx3
pyautogui
pywin32
fpdf
google-generativeai
